# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WAGE - API BI",
        "version": "0.1"
    },
    "paths": {
...
        "/api/usuarios": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Retorna a lista de usuarios",
                "description": "Retorna a lista de usuarios, usando como parametro de pesquisa o cnpj_empresa / email / nome / codigo_usuario ",
                "operationId": "userIndex",
                "parameters": [
                    {
                        "name": "cnpj_empresa",
                        "in": "query",
                        "description": "Cnpj da empresa ao qual o usuario pertence - orbigatorio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Usado para pesquisa quando somente este campos esta preenchido",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "Usado para pesquisa quando somente este campos esta preenchido",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "codigo_usuario",
                        "in": "query",
                        "description": "Usado para pesquisa quando somente este campos esta preenchido",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna a lista de empresas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuariosWithDate"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "login nao autorizado"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Insere o usuario no sistema",
                "description": "",
                "operationId": "userStore",
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/Usuarios"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuariosWithDate"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "login nao autorizado"
                    }
                }
            }
        },
        "/api/usuarios/{id}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Retorna o usuario",
                "description": "Retorna ousuario, usando como parametro o id",
                "operationId": "userShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuariosWithDate"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "login nao autorizado"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Remove o usuario",
                "description": "Remove o usuario selecionado, usando como parametro o id",
                "operationId": "userDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario excluido com sucesso"
                    },
                    "404": {
                        "description": "Usuario nao encontrado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/usuario/{id}": {
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Edita o usuario",
                "description": "Edita o usuario selecionado, usando como parametro o id",
                "operationId": "userUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/Usuarios"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o usuario alterado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuariosWithDate"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario nao encontrado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Usuarios": {
                "properties": {
                    "id": {
                        "description": "ID do usuario",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "codigo_usuario": {
                        "description": "Codigo do usuario no alfa",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "codigo_vendedor": {
                        "description": "Codigo do vendedor no alfa",
                        "type": "integer",
                        "example": "2",
                        "nullable": true
                    },
                    "nome": {
                        "description": "Nome usuario no alfa",
                        "type": "string",
                        "example": "Usuario Teste",
                        "nullable": false
                    },
                    "email": {
                        "description": "Email usado para realizar login",
                        "type": "string",
                        "example": "teste@teste.com",
                        "nullable": false
                    },
                    "password": {
                        "description": "Senha de acesso a api",
                        "type": "string",
                        "example": "123456",
                        "nullable": false
                    },
                    "senha_transacao": {
                        "description": "Senha usada para transacoes na api, como liberacao remota",
                        "type": "string",
                        "example": "654321",
                        "nullable": true
                    },
                    "ultimo_login": {
                        "description": "Data do ultimo login no sistema",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "gerente": {
                        "description": "Indica se o usuario e gerente",
                        "type": "string",
                        "example": "N",
                        "nullable": true
                    },
                    "vendedor": {
                        "description": "Indica se o usuario e vendedor",
                        "type": "string",
                        "example": "S",
                        "nullable": true
                    },
                    "recebe_notificacao": {
                        "description": "Indica se o usuario ira receber notificacao",
                        "type": "string",
                        "example": "N",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UsuariosWithDate": {
                "properties": {
                    "id": {
                        "description": "ID do usuario",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "codigo_usuario": {
                        "description": "Codigo do usuario no alfa",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "codigo_vendedor": {
                        "description": "Codigo do vendedor no alfa",
                        "type": "integer",
                        "example": "2",
                        "nullable": true
                    },
                    "nome": {
                        "description": "Nome usuario no alfa",
                        "type": "string",
                        "example": "Usuario Teste",
                        "nullable": false
                    },
                    "email": {
                        "description": "Email usado para realizar login",
                        "type": "string",
                        "example": "teste@teste.com",
                        "nullable": false
                    },
                    "password": {
                        "description": "Senha de acesso a api",
                        "type": "string",
                        "example": "123456",
                        "nullable": false
                    },
                    "senha_transacao": {
                        "description": "Senha usada para transacoes na api, como liberacao remota",
                        "type": "string",
                        "example": "654321",
                        "nullable": true
                    },
                    "ultimo_login": {
                        "description": "Data do ultimo login no sistema",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "gerente": {
                        "description": "Indica se o usuario e gerente",
                        "type": "string",
                        "example": "N",
                        "nullable": true
                    },
                    "vendedor": {
                        "description": "Indica se o usuario e vendedor",
                        "type": "string",
                        "example": "S",
                        "nullable": true
                    },
                    "recebe_notificacao": {
                        "description": "Indica se o usuario ira receber notificacao",
                        "type": "string",
                        "example": "N",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Data da criacao da empresa",
                        "type": "string",
                        "format": "date-time",
                        "nullable": false
                    },
                    "updated_at": {
                        "description": "Data da ultima alteracao da empresa",
                        "type": "string",
                        "format": "date-time",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}